Max out agent-native dictation hardening
